cellulitis-app/requirements.txt
flask
opencv-python-headless
numpy
scikit-image
pillow